requirements.txt
python-docx
pandas
openpyxl
regex
spacy
